+++ config/config.go
@@ -61,6 +61,13 @@ const (
 	// How many bytes per request
 	DefaultMaxBytesPerRequest = 65536
 
+	// How many items can be queued before the reader blocks
+	//
+	// Note: queue entries are 16 bytes, this is a large block of memory.
+	// TODO: Consider adjusting this down after understanding performance
+	// of the single-queue approach to sharding taken in #247.
+	DefaultQueueSize = 100000
+
 	// Min number of shards, i.e. amount of concurrency
 	DefaultMinShards = 1
 	// Max number of shards, i.e. amount of concurrency
@@ -200,16 +207,17 @@ type LogConfig struct {
 }
 
 type PromConfig struct {
-	Endpoint           string         `json:"endpoint"`
-	WAL                string         `json:"wal"`
-	MaxPointAge        DurationConfig `json:"max_point_age"`
-	MinShards          int            `json:"min_shards"`
-	MaxShards          int            `json:"max_shards"`
+	Endpoint    string         `json:"endpoint"`
+	WAL         string         `json:"wal"`
+	MaxPointAge DurationConfig `json:"max_point_age"`
 }
 
 type OTelConfig struct {
-	MaxBytesPerRequest int            `json:"max_bytes_per_request"`
-	MetricsPrefix string `json:"metrics_prefix"`
+	MaxBytesPerRequest int    `json:"max_bytes_per_request"`
+	MetricsPrefix      string `json:"metrics_prefix"`
+	MinShards          int    `json:"min_shards"`
+	MaxShards          int    `json:"max_shards"`
+	QueueSize          int    `json:"queue_size"`
 }
 
 type AdminConfig struct {
@@ -250,15 +258,11 @@ func (c MainConfig) QueueConfig() promconfig.QueueConfig {
 	cfg.MaxBackoff = model.Duration(2 * time.Second)
 	// Note: we are passing bytes in a MaxSamplesPerSend field.
 	cfg.MaxSamplesPerSend = c.OpenTelemetry.MaxBytesPerRequest
-	cfg.MinShards = c.Prometheus.MinShards
-	cfg.MaxShards = c.Prometheus.MaxShards
+	cfg.MinShards = c.OpenTelemetry.MinShards
+	cfg.MaxShards = c.OpenTelemetry.MaxShards
 
-	// We want the queues to have enough buffer to ensure consistent flow with full batches
-	// being available for every new request.
-	// Testing with different latencies and shard numbers have shown that 3x of the batch size
-	// works well.
-	// TODO: Use a single queue (in a future PR).
-	cfg.Capacity = 1500
+	// Note: This is size of a single queue owned by the queue manager.
+	cfg.Capacity = c.OpenTelemetry.QueueSize
 
 	return cfg
 }
@@ -268,14 +272,15 @@ type FileReadFunc func(filename string) ([]byte, error)
 func DefaultMainConfig() MainConfig {
 	return MainConfig{
 		Prometheus: PromConfig{
-			WAL:                DefaultWALDirectory,
-			Endpoint:           DefaultPrometheusEndpoint,
-			MaxPointAge:        DurationConfig{DefaultMaxPointAge},
-			MinShards:          DefaultMinShards,
-			MaxShards:          DefaultMaxShards,
+			WAL:         DefaultWALDirectory,
+			Endpoint:    DefaultPrometheusEndpoint,
+			MaxPointAge: DurationConfig{DefaultMaxPointAge},
 		},
 		OpenTelemetry: OTelConfig{
 			MaxBytesPerRequest: DefaultMaxBytesPerRequest,
+			MinShards:          DefaultMinShards,
+			MaxShards:          DefaultMaxShards,
+			QueueSize:          DefaultQueueSize,
 		},
 		Admin: AdminConfig{
 			Port:                      DefaultAdminPort,
@@ -353,12 +358,6 @@ func Configure(args []string, readFunc FileReadFunc) (MainConfig, map[string]str
 	a.Flag("prometheus.max-point-age", "Skip points older than this, to assist recovery. Default: "+DefaultMaxPointAge.String()).
 		DurationVar(&cfg.Prometheus.MaxPointAge.Duration)
 
-	a.Flag("prometheus.min-shards", fmt.Sprintf("Min number of shards, i.e. amount of concurrency. Default: %d", DefaultMinShards)).
-		IntVar(&cfg.Prometheus.MinShards)
-
-	a.Flag("prometheus.max-shards", fmt.Sprintf("Max number of shards, i.e. amount of concurrency. Default: %d", DefaultMaxShards)).
-		IntVar(&cfg.Prometheus.MaxShards)
-
 	var ignoredScrapeIntervals []string
 	a.Flag("prometheus.scrape-interval", "Ignored. This is inferred from the Prometheus via api/v1/status/config").
 		StringsVar(&ignoredScrapeIntervals)
@@ -374,8 +373,17 @@ func Configure(args []string, readFunc FileReadFunc) (MainConfig, map[string]str
 	a.Flag("opentelemetry.max-bytes-per-request", fmt.Sprintf("Send at most this many bytes per request. Default: %d", DefaultMaxBytesPerRequest)).
 		IntVar(&cfg.OpenTelemetry.MaxBytesPerRequest)
 
+	a.Flag("opentelemetry.min-shards", fmt.Sprintf("Min number of shards, i.e. amount of concurrency. Default: %d", DefaultMinShards)).
+		IntVar(&cfg.OpenTelemetry.MinShards)
+
+	a.Flag("opentelemetry.max-shards", fmt.Sprintf("Max number of shards, i.e. amount of concurrency. Default: %d", DefaultMaxShards)).
+		IntVar(&cfg.OpenTelemetry.MaxShards)
+
 	a.Flag("opentelemetry.metrics-prefix", "Customized prefix for exporter metrics. If not set, none will be used").
 		StringVar(&cfg.OpenTelemetry.MetricsPrefix)
+
+	a.Flag("opentelemetry.queue-size", fmt.Sprintf("Number of points that can accumulate before blocking the reader. Default: %d", DefaultQueueSize)).
+		IntVar(&cfg.OpenTelemetry.QueueSize)
 
 	a.Flag("filter", "PromQL metric and attribute matcher which must pass for a series to be forwarded to OpenTelemetry. If repeated, the series must pass any of the filter sets to be forwarded.").
 		StringsVar(&cfg.Filters)
@@ -448,7 +456,7 @@ func Configure(args []string, readFunc FileReadFunc) (MainConfig, map[string]str
 		}
 	}
 
-	if cfg.Prometheus.MinShards > cfg.Prometheus.MaxShards {
+	if cfg.OpenTelemetry.MinShards > cfg.OpenTelemetry.MaxShards {
 		return MainConfig{}, nil, nil, errors.New("min-shards cannot be greater than max-shards")
 	}