+++ config/config_test.go
@@ -146,7 +146,7 @@ func TestConfiguration(t *testing.T) {
 			"endpoint must be set: destination.endpoint",
 		},
 		{
-			"only_file", `
+			"only file", `
 destination:
   endpoint: http://womp.womp
   attributes:
@@ -160,7 +160,6 @@ destination:
 
 prometheus:
   wal: wal-eeee
-  scrape_intervals: [22m13s]
 
 startup_timeout: 1777s
 `,
@@ -172,11 +171,12 @@ startup_timeout: 1777s
 					MaxPointAge: DurationConfig{
 						25 * time.Hour,
 					},
-					MinShards: 1,
-					MaxShards: 200,
 				},
 				OpenTelemetry: OTelConfig{
 					MaxBytesPerRequest: 65536,
+					MinShards:          1,
+					MaxShards:          200,
+					QueueSize:          config.DefaultQueueSize,
 				},
 				Admin: AdminConfig{
 					ListenIP:                  config.DefaultAdminListenIP,
@@ -241,7 +241,7 @@ startup_timeout: 1777s
 		{
 			// Note that attributes and headers are merged, while
 			// for other fields flags overwrite file-config.
-			"file_and_flag", `
+			"file and flag", `
 destination:
   endpoint: http://womp.womp
   attributes:
@@ -268,8 +268,9 @@ log:
 				"--prometheus.wal", "wal-eeee",
 				"--prometheus.max-point-age", "10h",
 				"--opentelemetry.max-bytes-per-request", "5",
-				"--prometheus.min-shards", "5",
-				"--prometheus.max-shards", "10",
+				"--opentelemetry.min-shards", "5",
+				"--opentelemetry.max-shards", "10",
+				"--opentelemetry.queue-size", "107",
 				"--log.level=warning",
 				"--healthcheck.period=17s",
 				"--healthcheck.threshold-ratio=0.2",
@@ -285,11 +286,12 @@ log:
 					MaxPointAge: DurationConfig{
 						10 * time.Hour,
 					},
-					MinShards: 5,
-					MaxShards: 10,
 				},
 				OpenTelemetry: OTelConfig{
 					MaxBytesPerRequest: 5,
+					MinShards:          5,
+					MaxShards:          10,
+					QueueSize:          107,
 				},
 				Admin: AdminConfig{
 					ListenIP:                  config.DefaultAdminListenIP,
@@ -339,7 +341,7 @@ log:
 			"",
 		},
 		{
-			"all_settings", `
+			"all settings", `
 # Comments work!
 destination:
   endpoint: https://ingest.staging.lightstep.com:443
@@ -364,9 +366,6 @@ prometheus:
   wal: /volume/wal
   endpoint: http://127.0.0.1:19090/
   max_point_age: 72h
-  min_shards: 10
-  max_shards: 20
-  scrape_intervals: [30s]
 
 startup_timeout: 33s
 
@@ -387,7 +386,10 @@ security:
 
 opentelemetry:
   max_bytes_per_request: 10
+  min_shards: 10
+  max_shards: 20
   metrics_prefix: prefix.
+  queue_size: 701
 
 filters:
 - metric{label=value}
@@ -429,12 +431,13 @@ static_metadata:
 					MaxPointAge: DurationConfig{
 						72 * time.Hour,
 					},
-					MinShards: 10,
-					MaxShards: 20,
 				},
 				OpenTelemetry: OTelConfig{
 					MaxBytesPerRequest: 10,
 					MetricsPrefix:      "prefix.",
+					MinShards:          10,
+					MaxShards:          20,
+					QueueSize:          701,
 				},
 				Destination: OTLPConfig{
 					Endpoint: "https://ingest.staging.lightstep.com:443",
@@ -587,8 +590,8 @@ static_metadata:
 		{
 			"min-shards greater than max-shards", ``,
 			[]string{
-				"--prometheus.min-shards=101",
-				"--prometheus.max-shards=100",
+				"--opentelemetry.min-shards=101",
+				"--opentelemetry.max-shards=100",
 			},
 			config.MainConfig{},
 			"min-shards cannot be greater than max-shards",