+++ config/example_test.go
@@ -40,13 +40,14 @@ func Example() {
 	//   "prometheus": {
 	//     "endpoint": "http://127.0.0.1:19090",
 	//     "wal": "/volume/wal",
-	//     "max_point_age": "72h0m0s",
-	//     "min_shards": 100,
-	//     "max_shards": 200
+	//     "max_point_age": "72h0m0s"
 	//   },
 	//   "opentelemetry": {
 	//     "max_bytes_per_request": 1500,
-	//     "metrics_prefix": "prefix."
+	//     "metrics_prefix": "prefix.",
+	//     "min_shards": 100,
+	//     "max_shards": 200,
+	//     "queue_size": 100001
 	//   },
 	//   "admin": {
 	//     "listen_ip": "0.0.0.0",